backend/requirements.txt
psycopg2-binary==2.9.9
Faker==20.1.0
# Os itens abaixo são referentes ao roteiro
fastapi
uvicorn[standard]
sqlalchemy
asyncpg         # Driver assíncrono para PostgreSQL
pydantic
python-dotenv   # Para ler o arquivo.env
pytest
httpx